pipelinetest2 complete

+++ src/main/kaggle/classification/Sample.scala
@@ -1,16 +1,3 @@
 package main.kaggle.classification
 
-case class Sample(
-  PassengerId: Integer,
-  Survived: Integer,
-  Pclass: String,
-  Name: String,
-  Sex: String,
-  Age: Double,
-  SibSp: Double,
-  Parch: Double,
-  Ticket: String,
-  Fare: Double,
-  Cabin: String,
-  Embarked: String
-)
+case class Sample(PassengerId: Double, Survived: Double, Pclass: String, Name: String, Sex: String, Age: Double, SibSp: Double, Parch: Double, Ticket: String, Fare: Double, Cabin: String, Embarked: String)
